Update BankLogic.java
This took a bit of time to figure out, i was close to doing something really complicated.

But instead, i made something really simple. Elegant. I felt it's a smart solution to getting what Account number a Account should have, when a Account is created.

It also makes sure that we always get the Latest relevant number - No matter if there is a gap or if the Array is full of correct numbers already.

Simple, elegant, clever.